# specs/services/agent-research-orchestrator/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Research Orchestrator API
  description: REST API for the research orchestrator service, providing endpoints for managing research sessions and receiving real-time updates via SSE.
  version: 1.0.0
  contact:
    name: Platform Team
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://research.cofilot.demo
    description: Production
paths:
...
  /research/sessions:
    post:
      summary: Create a new research session
      operationId: createSession
      tags:
        - Research
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchRequest'
      responses:
        '201':
          description: Session created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchSession'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /research/sessions/{session_id}:
    get:
      summary: Get session status
      operationId: getSession
      tags:
        - Research
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchSession'
        '404':
          description: Session not found
  /research/sessions/{session_id}/start:
    post:
      summary: Start the research workflow
      operationId: startWorkflow
      tags:
        - Research
      parameters:
        - name: session_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Workflow started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowStarted'
        '404':
          description: Session not found
        '409':
          description: Workflow already running
...
components:
  schemas:
...
    ResearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          minLength: 10
          maxLength: 1000
          description: Research question
          example: "Should Cofilot expand to Vienna?"
        context:
          type: object
          additionalProperties: true
          description: Optional context for the research
    ResearchSession:
      type: object
      required:
        - id
        - query
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        query:
          type: string
        status:
          type: string
          enum: [created, running, waiting_input, completed, failed]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        current_agent:
          type: string
          nullable: true
        checklist:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        pending_questions:
          type: array
          items:
            $ref: '#/components/schemas/Question'
        final_report:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    ChecklistItem:
      type: object
      required:
        - id
        - task
        - agent
        - status
      properties:
        id:
          type: integer
        task:
          type: string
        agent:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, skipped]
        notes:
          type: string
          nullable: true
    Question:
      type: object
      required:
        - id
        - question
        - context
        - asked_by
        - priority
        - blocking
        - created_at
      properties:
        id:
          type: string
        question:
          type: string
        context:
          type: string
          description: Why this information is needed
        asked_by:
          type: string
          description: Agent that asked this question
        priority:
          type: string
          enum: [high, medium, low]
        blocking:
          type: boolean
          description: If true, workflow is paused waiting for this answer
        options:
          type: array
          items:
            type: string
          nullable: true
          description: Optional multiple choice options
        answer:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        answered_at:
          type: string
          format: date-time
          nullable: true
...
    WorkflowStarted:
      type: object
      required:
        - session_id
        - status
      properties:
        session_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [started]
        message:
          type: string
...
    ValidationError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string